backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
requests
PyPDF2
pydantic
